Validate 클래스 추가 및 메서드 생성을 통한 예외처리 정리
